improved construction of calc-results list from imca results
when inserting mission time in calc-result-item,
only use the mission time string from the imca result item,
when that mission time string != "" and != "?" --
otherwise, use the mission time string from the calcCommand.

This is meant to give us valid/useful missionTime in the calc-result
when the imca result does not contain a missiont time, which is
the case right now:
now, for each mission time we invoke imca separately,
and each imca result only contains a probability --  it does not
contain a mission time (yet).