Add toplevel process to remove pointless ACTIVATE self-loops.
We used to generate very Zeno processes where the top-level activation
could occur from every state (and usually do a self-loop). While this
generates the correct output after conversion to IMCA/MRMC, it's
inelegant and probably inefficient to have so many extra transitions.

We now have one template called 'toplevel' that fires one activation and
then does nothing. This was, we know the TLE is activated exactly once
at the beginning of the process.