Add verifyable remove of input signals.
Due to smart semantics, some automata can be substantially reduced by
omitting unneeded behaviour (e.g., responses to repairs in gates with
only non-repairable children). However, simply removing this may result
in incorrect behaviour if another element does not have the
corresponding semantics (e.g., a repairable BE will be permanently stuck
trying to emit its ONLINE signal if no gate is synchronising with it).

To signal such errors without adding unnecessary extra states, this
change adds an 'IMPOSSIBLE' signal. Every element may emit such a signal
after it received an input it does not expect. There is no requirement
on behaviour of the element after emitting this signal (my
recommendation is to deadlock, so that composing the element results in
a small model).

DFTCalc checks whether any IMPOSSIBLE signal is reachable, and errors
out if it is.