Fix handling of non-connected FDEP gates.
The FDEP gate itself has only a dummy output, so it does not need to be
attached to anything. However, activation of the trigger previously
only occurred if the FDEP gate was connected, which caused unexpected
results for trees containing such constructs.

Now, all parentless entities are labeled 'always active' which should
correct this issue.