Add caching of results.
Very frequently, subtrees (independent modules) will be identical modulo
naming. To avoid unnecessary recomputations, we cache the .exp file in
memory (they should be pretty small) along with the results.

This doesn't capture all cases of equivalent subtrees (in particular, it
will cases where children of boolean gates are reordered) but it already
saves a lot.