Simplified inspection module.
The inspection module now does a simultaneous repair on all children
rather than a nondeterministic one.

In the current semantics this should make no difference as there are no
gates that depend on simultaneous synchronization with multiple
children, so after repair the children will just nondeterministically
choose the order to communicate with their parents.

In the mean time, it should significantly reduce the intermediate
state-space as the number of interleavings is greatly reduced.